ResourceBased

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesCommandHandler.cs
@@ -2,18 +2,29 @@
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
+using Restaurants.Infrastructure.Authorization;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
 
 public class DeleteDishesCommandHandler(ILogger<DeleteDishesCommandHandler> logger ,
-    IRestaurantsRepository restaurantsRepository,IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesCommand>
+    IRestaurantsRepository restaurantsRepository,IDishesRepository dishesRepository, 
+    IRestaurantAuthorizationService restaurantAuthorizationService) 
+        : IRequestHandler<DeleteDishesCommand>
 {
     public async Task Handle(DeleteDishesCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation("Deleting All Dishes for Restaurant with id: {restaurant.Id}",request.RestaurantId);
+        
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId)
             ?? throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+        {
+            throw new ForbidException();
+        }
+
         await dishesRepository.DeleteAsync(restaurant.Dishes);
     }
 }
